Level 2 - Introduction to Software Engineering/M06T08 – Capstone Project – News Application/project_news/requirements.txt
alabaster==1.0.0
asgiref==3.11.1
astroid==4.0.4
babel==2.18.0
black==26.1.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
dill==0.4.1
Django==6.0.2
djangorestframework==3.16.1
docutils==0.22.4
idna==3.11
imagesize==1.4.1
isort==7.0.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mccabe==0.7.0
mypy_extensions==1.1.0
mysqlclient==2.2.8
oauthlib==3.3.1
packaging==26.0
pathspec==1.0.4
platformdirs==4.9.2
Pygments==2.19.2
pylint==4.0.4
pylint-django==2.7.0
pylint-plugin-utils==0.9.0
pytokens==0.4.1
requests==2.32.5
requests-oauthlib==2.0.0
roman-numerals==4.1.0
snowballstemmer==3.0.1
Sphinx==9.1.0
sphinx_rtd_theme==3.1.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.5
tomlkit==0.14.0
urllib3==2.6.3
